+++ Front-end/js/game/game.js
@@ -282,7 +282,7 @@ function showBackToLobbyBtn()
 
 function gameEnded(winner)
 {
-    window.alert("Team " + str(winner) + " won the game!");
+    setTimeout(() => window.alert("Team " + str(winner) + " won the game!"), 0);
 }
 
 function resetRoom()
@@ -408,38 +408,20 @@ class Tile
     {
         let wordObj = SERVER.room.game.words[this.index];
         this.team = wordObj.team;
-        if (this.team === -2) // Change -2 to null
+
+        if (this.team === -2)
         {
             this.elem.style.backgroundColor = "gray";
             this.elem.style.color = "var(--topColor)";
         }
-        else if (this.team === -1)
-        {
-            this.elem.style.backgroundColor = "var(--backColor)";
-            this.elem.style.color = "var(--topColor)";
-        }
-        else if (this.team === 0)
-        {
-            this.elem.style.backgroundColor = "var(--topColor)";
-            this.elem.style.color = "var(--backColor)";
-        }
         else
         {
             this.elem.style.backgroundColor = teams[this.team + 1].normal;
-            this.elem.style.color = getColorsForElem(this.elem).color;
+            setTimeout(() => this.elem.style.color = getColorsForElem(this.elem).color, 0);
         }
     }
 
-    select()
-    {
-        this.update();
-        this.selectedBy = SERVER.room.game.words[this.index].selectedBy;
-
-        this.elem.style.borderColor = teams[this.selectedBy + 1].light;
-        this.elem.style.backgroundColor = teams[this.team + 1].light;
-    }
-
-    mark()
+    updateMark()
     {
         this.marked = getMarked(SERVER.room, this.index);
 
@@ -462,4 +444,15 @@ class Tile
             this.update();
         }
     }
+
+    select()
+    {
+        this.updateMark();
+        this.update();
+        this.selectedBy = SERVER.room.game.words[this.index].selectedBy;
+
+        this.elem.style.borderColor = teams[this.selectedBy + 1].light;
+        this.elem.style.backgroundColor = teams[this.team + 1].light;
+    }
+
 }